+++ TccMeAdoteApi/src/main/java/com/tccmeadote/meadote/repositories/FavoritesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface FavoritesRepository extends JpaRepository<Favorites, Integer> {
-    Optional<Favorites> findByUserFirebaseUidAndPostId(String userFirebaseUid, Integer postId);
+public interface FavoritesRepository extends JpaRepository<Favorites, Long> {
+    Optional<Favorites> findByUserFirebaseUidAndPostId(String userFirebaseUid, Long postId);
 }
